Add a separate .js and .test.js file for the addTask function
