+++ src/__mocks__/addTask.js
@@ -1,0 +1,9 @@
+import TaskUtils from '../taskUtils.js';
+import displayTodo from './displayTodo.js';
+
+const taskUtils = new TaskUtils();
+export const tasks = taskUtils.getTasks();
+export default function addTask(task) {
+  tasks.push(task);
+  displayTodo(tasks);
+}
