+++ src/__tests__/taskUtils.test.js
@@ -2,50 +2,14 @@
  * @jest-environment jsdom
  */
 
-import TaskUtils from '../taskUtils.js';
+import addTask, { tasks } from '../__mocks__/addTask.js';
 
 jest.mock('../taskUtils.js');
-
-const taskUtils = new TaskUtils();
 
 document.body.innerHTML = '<div>'
     + '<ul id="list"></ul>'
     + '</div>';
-
-const displayTodo = (arr) => {
-  const ul = document.getElementById('list');
-  arr.forEach((e) => {
-    const li = document.createElement('li');
-    ul.appendChild(li);
-    const checkBox = document.createElement('input');
-    checkBox.type = 'checkbox';
-    const label = document.createElement('label');
-    label.textContent = e.description;
-    li.appendChild(checkBox);
-    li.appendChild(label);
-    ul.appendChild(li);
-  });
-};
-describe('add exactly one <li> element to the list in the DOM', () => {
-  const task = { description: 'Task 1', completed: false, index: 1 };
-  taskUtils.addTask(task);
-  const tasks = taskUtils.getTasks();
-  displayTodo(tasks);
-
-  test('that the length of the list is 1', () => {
-    const list = document.querySelectorAll('#list li');
-    expect(list).toHaveLength(1);
-  });
-
-  test('that the length of the list is not zero', () => {
-    const list = document.querySelectorAll('#list li');
-    expect(list).not.toHaveLength(0);
-  });
-
-  test('that the description is Task 1', () => {
-    expect(tasks[0].description).toEqual('Task 1');
-  });
-});
+const task = { description: 'Task 1', completed: false, index: 1 };
 
 describe('Edit one <li> element to the list in the DOM', () => {
   test('that the input of the list is editing', () => {
